ComplexityPipeline/requirements.txt
torch>=2.1.0,<2.6
numpy>=1.25.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
watchdog>=3.0.0
prefect>=2.14.0
pymongo>=4.6.0
PyYAML>=6.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.66.0
shap>=0.44.0
streamlit>=1.30.0
pydantic>=2.5.0
prometheus-fastapi-instrumentator>=6.0.0
psutil>=5.9.0
redis>=4.6.0

# --- Dev / Test Dependencies ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.4